Fix momentary black screen when exiting full screen mode, #5600
This commit will:
- Add suspendDrawingNewFrames and resumeDrawingNewFrames
  resumeDrawingNewFrames to ViewLayer to provide the ability to suspend
  drawing of new video frames from mpv
- Add similar methods to ViewView that delegate to ViewLayer
- Add an isSuspended property to ViewLayer
- Add a updateFromMPV function to ViewLayer
- Change MVPController.mpvUpdateCallback to call updateFromMPV
- Remove setting of inLiveResize in the shadow copy initializer to
  correct a hang
- Correct checking for forced drawing in canDraw
- Change forced drawing back to using the main thread
- Change MainWindowController.toggleWindowFullScreen to suspend drawing
  of new video frames from mpv before starting to enter or exit full
  screen mode
- Change legacyAnimateToFullscreen and legacyAnimateToWindowed to wrap
  operations in a NSAnimationContext
- Change windowDidEndLiveResize to not call updateWindowParametersForMPV
  when entering or exiting full screen mode

This actually does not fix the momentary black screen that occurs when
entering or exiting full screen mode when using legacy full screen mode.
The black screen is caused by AppKit erasing the window and then drawing
a black background. The changes order operations to minimize the time a
black window is displayed before the views are drawn in order to make
the erasing of the window imperceptible.

This also corrects problems with the floating OSC being malformed after
entering / exiting full screen mode. And a hang when attempting to apply
a custom crop when in full screen mode.

These changes remove the fix for issue #5352, "Audio Stutter and Dropped
Frames When Entering Full-Screen Mode". The fix for that issue allowed
background drawing to continue while entering and exiting full screen
mode. This works some of the time, but frequently the main thread hangs
for enough time to noticeably disrupt animations. This bad behavior is
more apparent to the user than the dropped frames.